+++ pkg/servers/everything/log.go
@@ -3,16 +3,16 @@ package everything
 import "github.com/MegaGrindStone/go-mcp/pkg/mcp"
 
 // LogStreams implements mcp.LogHandler interface.
-func (s *SSEServer) LogStreams() <-chan mcp.LogParams {
+func (s *Server) LogStreams() <-chan mcp.LogParams {
 	return s.logChan
 }
 
 // SetLogLevel implements mcp.LogHandler interface.
-func (s *SSEServer) SetLogLevel(level mcp.LogLevel) {
+func (s *Server) SetLogLevel(level mcp.LogLevel) {
 	s.logLevel = level
 }
 
-func (s *SSEServer) log(msg string, level mcp.LogLevel) {
+func (s *Server) log(msg string, level mcp.LogLevel) {
 	if level < s.logLevel {
 		return
 	}