+++ pkg/servers/everything/progress.go
@@ -3,6 +3,6 @@ package everything
 import "github.com/MegaGrindStone/go-mcp/pkg/mcp"
 
 // ProgressReports implements mcp.ProgressReporter interface.
-func (s *SSEServer) ProgressReports() <-chan mcp.ProgressParams {
+func (s *Server) ProgressReports() <-chan mcp.ProgressParams {
 	return s.progressChan
 }