+++ pkg/servers/everything/prompt.go
@@ -45,7 +45,7 @@ var promptCompletions = map[string][]string{
 }
 
 // ListPrompts implements mcp.PromptServer interface.
-func (s *SSEServer) ListPrompts(_ context.Context, _ mcp.PromptsListParams) (mcp.PromptList, error) {
+func (s *Server) ListPrompts(context.Context, mcp.PromptsListParams, mcp.RequestClientFunc) (mcp.PromptList, error) {
 	s.log("ListPrompts", mcp.LogLevelDebug)
 
 	s.log(fmt.Sprintf("count prompts: %d", len(promptList.Prompts)), mcp.LogLevelInfo)
@@ -54,7 +54,11 @@ func (s *SSEServer) ListPrompts(_ context.Context, _ mcp.PromptsListParams) (mcp
 }
 
 // GetPrompt implements mcp.PromptServer interface.
-func (s *SSEServer) GetPrompt(_ context.Context, params mcp.PromptsGetParams) (mcp.PromptResult, error) {
+func (s *Server) GetPrompt(
+	_ context.Context,
+	params mcp.PromptsGetParams,
+	_ mcp.RequestClientFunc,
+) (mcp.PromptResult, error) {
 	s.log(fmt.Sprintf("GetPrompt: %s", params.Name), mcp.LogLevelDebug)
 
 	switch params.Name {
@@ -120,9 +124,10 @@ I understand. You've provided a complex prompt with temperature and style argume
 }
 
 // CompletesPrompt implements mcp.PromptServer interface.
-func (s *SSEServer) CompletesPrompt(
+func (s *Server) CompletesPrompt(
 	_ context.Context,
 	params mcp.CompletionCompleteParams,
+	_ mcp.RequestClientFunc,
 ) (mcp.CompletionResult, error) {
 	s.log(fmt.Sprintf("CompletesPrompt: %s", params.Ref.Name), mcp.LogLevelDebug)